backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.4
defusedxml==0.7.1
Django==3.2.16
django-filter==23.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
flake8==6.1.0
flake8-django==1.2
flake8-isort==6.0.0
flake8-quotes==3.4.0
idna==3.10
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
oauthlib==3.3.1
packaging==25.0
pillow==11.2.1
pipdeptree==2.28.0
psycopg2-binary==2.9.3
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.1.0
PyJWT==2.10.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
webcolors==1.11.1
